dags/scraper/requirements.txt
async-generator==1.10
attrs==22.2.0
azure-core==1.26.2
azure-storage-blob==12.14.1
beautifulsoup4==4.11.2
bs2json==0.0.3
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
collection==0.1.6
cryptography==39.0.0
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
isodate==0.6.1
joblib==1.2.0
msrest==0.7.1
oauthlib==3.2.2
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
requests==2.28.2
requests-oauthlib==1.3.1
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.14
wsproto==1.2.0
